feature: add a generic deploy function with an evm implementation
TODO
- Update changeset when implemented
- Add tests for EVM implementation
- Consider how we want to structure chain family specific implementations beyond just EVM

Adds a new package `contract` containing a generic `Deploy`` function
which accepts a `DeployFunc` function to perform the deployment of a
contract.

Specific chain families can provide implementations of the `DeployFunc`
for each of the contracts that are required to be deployed for their
on-chain implementation of MCMS. A implementation has been added for
EVM chains using the `go-ethereum`` library.

```
import (
    "github.com/smartcontractkit/mcms/pkg/contract"
    "github.com/smartcontractkit/mcms/sdk/evm"
)

// Ethereum setup for auth and backend...

contract.Deploy(evm.MCMSContractDeployment(auth, backend))
````