chore: setup lint runs in CI and add more linters
Adds a CI job to run the linters when a PR is opened against main, or
is merged to main. More linters have also been added to the project to
ensure that the codebase is of high quality. Some linters have been
disabled as they currently fail and will be enabled in follow up PRs.

This also updates the linting taskfile to use the default task so that
users can run `task lint` instead of `task lint:check`.