chore: moves operation and metadata types into a types package (#81)
In order to remove the potential for circular dependencies, the
operation and metadata types have been moved into a types package. This
will allow for the types to be used in multiple packages without causing
import cycles and can easily be shared between the main core code, and
the sdk packages.