chore: rename TimelockOperationType to TimelockAction and move to the types package (#82)
`TimelockOperationType` has been renamed to `TimelockAction` to avoid
confusion with an operation.

This has also been moved to the `types` package, which led to an issue
that could cause potential naming conflicts due to lack of context on
the name of the const (types.Schedule, types.Cancel, etc). This has now
been renamed to prefix `TimelockAction` to each action.