chore: move `mcms.NewEncoder` to the sdk package (#96)
Having `NewEncoder` in the `mcms` package leaks chain specific details
into the `mcms` package, as it is required to determine the chain family
of the selector to build the correct sdk encoder. To solve this, we can
move the `NewEncoder` function to the `sdk` package, which isolates all
chain specific details to the `sdk` package.