chore: refactor mcms simulated backend tests (#98)
This adds a new package `internal/testutils/evmsim` which provides a
wrapper around the go-ethereum simulated backend. It provides a
`SimulatedBackend` struct which exposes methods to deploy and call MCM
related contracts. This splits up the previous single function setup
call so it gives the caller more control over how they want to build
their tests.

This can later be extended to add more common methods for interacting
with MCM contracts on the simulated backend as we expand the test suite.