chore: refactor NewExecutable to be the main entry point to creating an executor (#101)
This removes the `ToExecutable` method on the proposal, and changes
`NewExecutable` to accept a proposal and return an executor. This aligns
more closely with go paradigms to inject dependencies into a struct,
rather than the more OO approach of having a method on a struct that
returns a new struct.