chore: change public Signing interface (#102)
This changes the public Signing interface to include a Sign method that
accepts a Signable object, and a Signer implementation and returns the
signature.

This change is necessary to allow for different signing algorithms to be
used in the future, while reducing the duplication of the code.

```
signable, err := proposal.Signable(true, inspectors)

Sign(signable, NewPrivateKeySigner(sim.Signers[0].PrivateKey))
 // OR
Sign(signable, NewLedgerSigner([]uint32{1,1,1,1}))
````