feat: separate base proposal and use validator pkg for basic validations [DPA-1214] (#103)
* The JSON validation logic for the timelock struct was a bit convoluted
due to the duplicate `Transactions` field which was overridden from the
`MCMSProposal`. This PR aims to simplify the data structure by
separating the shared fields into a `BaseProposal` to simplify the
marshal/unmarshal process.
* We also add some go tags to include basic validations leveraging the
validator pkg. This helps us give more granular errors on the specific
fields that failed, as well as removing some error types that are not
handled by the validator package.