feat: using EVM simulated chains is now defined on a proposal (#105)
A new private field `useSimulatedBackend` has been added to the
MCMSProposal struct, which is used to determine whether the proposal
should be executed on a simulated chain. For EVM, this overwrites the
EVMChainID, however every chain family sdk may implement this
differently.

This does not aim to modify the JSON representation of the proposal.
Instead, usage of simulated backends will be done by calling the
`UseSimulatedBackend` function on the proposal after marshalling the
JSON.

This allows us to avoid passing in an `isSim` boolean around to multiple
functions, so that we can read it from the proposal itself.