chore: move MerkleTree generation to the Proposal
Move the MerkleTree generation to the Proposal struct to make it easier
to share between the Executable and Signable structs.

Instead of have a single loop on struct init that generates the
MerkleTree and determines the nonce values of the transactions, two
functions have been added to the proposal which provide these values.
While this does result in a slight calculation overhead, this is has
been mitigated by caching the resulting values.