chore: move MerkleTree generation to the Proposal (#112)
Move the MerkleTree generation to the Proposal struct to make it easier
to share between the Executable and Signable structs.

Instead of have a single loop on struct init that generates the
MerkleTree and determines the nonce values of the transactions, two
functions have been added to the proposal which provide these values.