chore: add workflow to mark PRs as stale after 60 days
The workflow is triggered daily at midnight UTC.

A PR with more than 60 days of inactivity will be marked as stale.

A PR that's stale for more than 7 days will be automatically closed.

Issues are exempt from auto marking as stale but issues with manually
added 'stale' label are eligible for auto closure after 7 days.

PRs with assignees are exempt from auto stale marking, it's the
responsibility of the assignee to get the PR progressed either with
review/merge or closure.