feat: changes the data structure of operations and it's use in proposals
This is a large change that changes the data structure of operations.
The current implementation promotes confusion with the usage of terms
operations and transactions. This change aims to make the codebase
more readable and maintainable, and provide a clear distinction between
operations and transactions.

Operations are now a separate entity from transactions. Operations are
defined as a piece of work that the MCMS can execute, while transactions
are the low level chain specific data structure that the MCMS can hash,
sign and execute on chain.