refactor: removes the EVM Chain ID argument from newEncoder (#144)
`newEncoder` was leaking chain family specific information into the
`mcms` package as we had to fetch the EVM chain id to construct an EVM
encoder.

This change removes the EVM chain id argument from `newEncoder` and
replaces it with the chain selector, so that the sdk itself is
responsible for fetching it's chain family specific chain id.

This also allows us to remove the `isSim` override from the `mcms` into
the evm sdk itself, which is a better place for it since this is evm
specific.