chore: use builder API on unit tests (#150)
Refactors some tests to use the builder API. Not all tests can/should be
using this API, for example validation tests where left the same as we
need the flexibility to deliberately leave some fields blank or with
incorrect values to trigger the errors. But on the tests that it was
possible the change was made.