docs(docsify): auto scroll up on page change
Currently there is an annoying behaviour where if you change pages, it does not reset the scroll bar to the top.
If you are viewing the bottom of a page and change to a new change, scrollbar remains at the bottom of the page and you have to scroll up manually.

Setting auto2top will reset scrollbar to the top on page change.
JD has this set to true as well.