tests/ledger signing (#194)
By Default, the `SignText` function in the geth usbwallet package
returns NotSupported. To resolve this, we fork that package and enable
that function to be used and use that here.

Also, this PR also updates some of the `Sign` logic in `Signable`:
* `SigningHash()` always returns the hash with the EIP 191 prefix
* `SigningMessage()` only returns the hash without the EIP 191 prefix
* The `Sign` method in `Signable` is updated to always use
`SigningMessage` instead of `SigningHash` (this is because the EIP 191
prefix is automatically added in the ledger case)
* The `Sign` function on the private key signer is updated to always add
an EIP 191 prefix before signing

This fixes two issues:
* The `NotSupported` error when using ledger
* The mismatching recovered address in the ledger E2E test

---------

Co-authored-by: ajaskolski <adrian.jaskolski@smartcontract.com>
Co-authored-by: ajaskolski <34792957+ajaskolski@users.noreply.github.com>