tests(solana): simple setup mcms e2e for solana
Still need more refactoring, but wanted to get this e2e tests in as an example of deploying MCMS using CTF framework to a solana node.
To start, i am commiting the mcms rust compiled files into the repo, we can think about dynamically creating it in the future. This is just to get things going.

JIRA: https://smartcontract-it.atlassian.net/browse/DPA-1413