[DPA-1413]: tests(solana): simple setup mcms e2e for solana (#205)
Still need more refactoring, but wanted to get this e2e tests in as an
example of deploying MCMS using CTF framework to a solana node.

I am piggy-bagging on a bunch of functions defined in `chainlink-ccip`
as they already have working test for MCMS, i just ported a simple one
over and modified it to use CTF as example.

JIRA: https://smartcontract-it.atlassian.net/browse/DPA-1413

This change relies on this CTF PR being merged
https://github.com/smartcontractkit/chainlink-testing-framework/pull/1519