test(solana): reset program id and remove SetProgramID for non write operations (#243)
I notice currently we are not testing the SetProgramID properly, if we
miss [the line to SetProgramID
](https://github.com/smartcontractkit/mcms/blob/29115f85234eeacc920e272f02f6be832bbcce90/sdk/solana/configurer.go#L54-L54)in
our methods, our tests still pass because we set this in the initial
`setup` for e2e, unit tests will pass because we mock the solana client.

Reset program Id for each program before each e2e test is run.

This ensures all methods in the sdk correctly sets the programID by
itself and not rely on the global programID to be set by something else

I verified this behavior when i commented out the SetProgramID in the
SetConfig method, and the test fails with
```
 "Transaction simulation failed: Attempt to load a program that does not exist",
```

--- 

After some testing and digging at the code, seems like SetProgramID is
only required for write operations and not needed for read operations as
proven by the e2e tests.