feat!: add a context and Converter map params to TimelockConverter.Convert
Add two parameters to the `TimelockProposal.Convert()` interface:
```go
func (m *TimelockProposal) Convert(
	ctx context.Context,
	converters map[types.ChainSelector]sdk.TimelockConverter,
) (Proposal, []common.Hash, error) {
  // ...
}
```

### Why

The Solana sdk requires us to access the blockchain to fetch the
configuration of the given timelock contract. In order to access the
blockchain, we need a `context` and an sdk-specific conveter instance
(which, in solana's case, holds a reference to an rpc client).

BREAKING_CHANGE