feat!: add context and Converter map to TimelockConverter.Convert params (#245)
This PR contains to main changes to the "convert timelock proposal"
workflow and related types and APIs:

1. Add two parameters to the `TimelockProposal.Convert()` interface:
```go
func (m *TimelockProposal) Convert(
	ctx context.Context,
	converters map[types.ChainSelector]sdk.TimelockConverter,
) (Proposal, []common.Hash, error) {
  // ...
}
```

2. Modify `sdk.TimelockConverter.ConvertBatchToChainOperation` adding a
`context` parameter and returning a slice of `types.Operation` instead
of just one:
```go
type TimelockConverter interface {
	ConvertBatchToChainOperations(
		ctx context.Context,
		bop types.BatchOperation,
		timelockAddress string,
		delay types.Duration,
		action types.TimelockAction,
		predecessor common.Hash,
		salt common.Hash,
	) ([]types.Operation, common.Hash, error)
}
```

### Why

The Solana sdk requires us to access the blockchain to fetch the
configuration of the given timelock contract. In order to access the
blockchain, we need a `context` and an sdk-specific converter instance
(which, in solana's case, holds a reference to an rpc client).

Additionally, the Solana version of the timelock contract might require
multiple instructions to schedule a "timelock operation" (not to be
confused with the `types.Operation`):
* `initialize`
* `append` (one or more)
* `finalize`
* `scheduleBatch`

So the `ConvertBatchToChainOperation` now needs to return multiple
`types.Operation`. It was renamed to `ConvertBatchToChainOperations` to
reflect that.

---
BREAKING_CHANGE