feat: alternative design for Solana's Simulator (#254)
Add `sdk.solana.Simulator` type, which allows one to simulate solana
transactions. For now, only the SetRoot operation is fully implemented
and tested.

In order to implement the simulator, the following changes were applied
to the solana SDK:

* `sendAndConfirm` is not parameterized using generics anymore. This is
to use it as a parameter. The downside is that now we don't check the
type of the instruction builders at compile type, we do it at runtime
(and new builders must be added explicitly).
* `sendAndConfirm` is now a parameter of the `Executor` type. It
defaults to the standard implementation and may be overwritten using the
`withSendAndConfirm` method
* There's a new `Simulator.collectInstructions` function with the same
signature as `sendAndConfirm`. It saves to built instructions in memory.
The simulator type replaces the default `sendAndConfirm` with
`collectInstructions` when constructed
* The simulator also has a `Simulator.simulate` function which simulates
the collected instruction. It's called after calling
`executor.SimulatorSetRoot`.

`SimulateOperation` has an initial implementation, but it still needs to
be properly tested. This will be done in a follow up ticket.

JIRA: https://smartcontract-it.atlassian.net/browse/DPA-1417

---------

Co-authored-by: Graham Goh <graham.goh@smartcontract.com>