feat: improve ledger testing for solana (#272)
Improves ledger testing by adding calls to SetRoot and including a
solana tx in the proposal. The SetRoot calls ensures that the signature
was accepted by the contract and not only that the signature was
appended to the proposal.

## AI Summary
This pull request includes several changes to the `e2e` and `sdk/solana`
directories, focusing on improving the handling of Solana and EVM chains
in the MCMS (Many Chain Multi-Sig) system. The most important changes
include adding support for Solana transactions in end-to-end tests,
refactoring the initialization of the MCM program, and updating the
executor creation for Solana chains.

### Support for Solana Transactions in E2E Tests:
* Added a new Solana transaction to the `proposal-testing.json` fixture
to support end-to-end tests involving Solana chains.
* Updated the `ledger_test.go` file to include Solana-specific imports
and added a new test suite `ManualLedgerSigningTestSuite` to test manual
ledger signing functionality for both EVM and Solana chains.
[[1]](diffhunk://#diff-046f7b9890699bc5c0782f4c739ef15b5f8be4727b0c2f7cf932fa171d9f514dL1-R158)
[[2]](diffhunk://#diff-046f7b9890699bc5c0782f4c739ef15b5f8be4727b0c2f7cf932fa171d9f514dL67-R237)
[[3]](diffhunk://#diff-046f7b9890699bc5c0782f4c739ef15b5f8be4727b0c2f7cf932fa171d9f514dR263-R266)

### Refactoring Initialization of MCM Program:
* Refactored the `SetupMCM` function to a public helper function
`InitializeMCMProgram` in `common.go` to initialize the MCM program with
the given PDA seed. This change was applied across multiple test files
to standardize the initialization process.
[[1]](diffhunk://#diff-08c1332aeae8454816329010fb4af1c91528b8059a04739329f4cdf8218f06d0L133-R184)
[[2]](diffhunk://#diff-e1fbbd2697b9af2cb0cd1138b6b678e36a35ed1271f94c24fe9489d1ed159cebL32-R32)
[[3]](diffhunk://#diff-d45f8b1fc9ce32967491b3200326fd5fde2a829bc865c8e52cf6a09778fcc5f8L20-R20)
[[4]](diffhunk://#diff-7e3c7e759e7d2c3b7900d53fd6411bbcbcead44a2d1e6ca7b14b85f32574bc64L21-R21)
[[5]](diffhunk://#diff-9b550f2222fbbcfc8af5b5ed60c0b4567f815d9aeb1aecd54ab08b864fd483c0L25-R25)
[[6]](diffhunk://#diff-715e94216372225cfc4c7393c82b553f0b784dcd30d21f067bade69b1ef7c38cL42-R42)

### Updating Executor Creation for Solana Chains:
* Modified the `NewExecutor` function in `executor.go` to change the
order of parameters, placing the encoder parameter first. Corresponding
changes were made in the relevant test files to reflect this update.
[[1]](diffhunk://#diff-2890f328acff17cf8abc91d22a143ac8a1d955b60ec675f6c57770df1d20f860L33-R33)
[[2]](diffhunk://#diff-350968eef1e410e9af2dba47013590938a30c4391bed93d2329c376a33118331L32-R32)
[[3]](diffhunk://#diff-350968eef1e410e9af2dba47013590938a30c4391bed93d2329c376a33118331L173-R173)

These changes enhance the integration and testing capabilities for
Solana chains within the MCMS system, ensuring better support for
multi-chain operations.