fix(solana): fix simulator side effect bug
If we create a simulator with an executor, on construction of that simulator, the sendAndConfirm function is overwritten with a no-op version.

```
executor := mcmsSolana.NewExecutor(s.SolanaClient, auth, encoder)
simulator := mcmsSolana.NewSimulator(executor)
simulator.SimulateSetRoot(....)

executor.SetRoot(....) ????? This will not perform any transaction
```

This commit ensure a copy of executor is modified.

JIRA: https://smartcontract-it.atlassian.net/browse/DPA-1417