fix: chain/operation timelock isReady helpers (#295)
This pull request introduces additional helper functions to the
`TimelockExecutable` API to enhance readiness checks. The main changes
include the addition of `IsChainReady` and `IsOperationReady` methods,
as well as updates to the existing `IsReady` method to incorporate these
new checks. Additionally, new tests have been added to verify the
functionality of these methods.

Enhancements to readiness checks:

* Added `IsChainReady` method to `timelock_executable.go` to check if
the chain is ready for execution.
* Added `IsOperationReady` method to `timelock_executable.go` to check
if a specific operation is ready.
* Updated `IsReady` method to call `IsOperationReady` for each global
operation in the proposal.
[[1]](diffhunk://#diff-5a1d103cef63b5dba0cfe2c93618b65a233c43862c69a6ba89fbee2b09bbc1ceR83-R123)
[[2]](diffhunk://#diff-5a1d103cef63b5dba0cfe2c93618b65a233c43862c69a6ba89fbee2b09bbc1ceL98-R133)

Testing improvements:

* Added tests for `IsChainReady` method in `timelock_executable_test.go`
to ensure it fails and succeeds under appropriate conditions.
[[1]](diffhunk://#diff-2ee0863eda6e2cdffe4fda89ec41b6a744cf57c0028ffba28c5d2b7bab8cbc55R549-R554)
[[2]](diffhunk://#diff-2ee0863eda6e2cdffe4fda89ec41b6a744cf57c0028ffba28c5d2b7bab8cbc55R563-R568)

Documentation updates:

* Updated `.changeset/clean-turtles-peel.md` to document the addition of
new `isReady` helpers to the API.