fix: make predecessors parameter optional
Predecessors can be optional and user should not be forced to provide an empty value if not needed on the API for NewProposal and NewTimelockProposal.
This commit introduces a functional option for `predecessors` and making that config optional, so the api is cleaner.

This is a breaking change, but better do it now than later before we have users.