fix: make predecessors parameter optional (#297)
Predecessors can be optional and user should not be forced to provide an
empty value if not needed on the API for `NewProposal` and
`NewTimelockProposal`. This commit introduces a functional option for
`predecessors` and making that config optional, so the api is cleaner.

This change was created after a brief discussion with James.
This is a breaking change, so open to recommendations if we want to do
it or not ( do we care enough?) but better do it now than later before
we have more users.

This is our official first breaking change, this will be a major version
bump.