feat: proposal decoding interfaces + EVM implementation (#301)
This pull request introduces a new interface for decoding operations and
implements it for the EVM decoder. The most important changes include
the creation of a `DecodedOperation` interface, updates to the `Decoder`
interface, and the implementation of these interfaces in the EVM
decoder.

### Interface and Type Definitions:

* Added `DecodedOperation` interface with methods `MethodName`, `Args`,
and `String` in `sdk/decoder.go`.
* Updated `Decoder` interface to use `DecodedOperation` as the return
type of the `Decode` method in `sdk/decoder.go`.

### EVM Decoder Implementation:

* Implemented `DecodedOperation` struct with methods `MethodName`,
`Args`, and `String` in `sdk/evm/decoder.go`.
* Implemented `Decoder` struct with the `Decode` method in
`sdk/evm/decoder.go`.
* Added `ParseFunctionCall` function to parse the ABI and transaction
data into a `DecodedOperation` in `sdk/evm/decoder.go`.

---------

Co-authored-by: Graham Goh <graham.goh@smartcontract.com>