[WIP] feat(evm): add support for simulating multiple calls atomically
We rely on the new `evm_simulateV1` json rpc endpoint to do this.
Support is still spotty as it was added to geth in 1.14.9 and most RPCs
do not seem to provide it yet. The implementation is also a bit low
level as the geth Go bindings do not provide a wrapper to this endpoint.