feat: add IsOperationDone and IsChainDone to timelock executable (#337)
In order to support concurrent timelock executions between the
chainlink-deployments CI and the legacy timelock-worker service, the GHA
running on the CI need to be able to check if an operation and/or the
operations of a chain have been marked as "done" by the timelock worker.

This PR adds two methods to the TimelockExecutable type, implementing
these two checks. The implementation was adapted from the existing
methods: `IsOperationReady` and `IsChainReady`.

JIRA: [DPA-1613](https://smartcontract-it.atlassian.net/browse/DPA-1613)

[DPA-1613]:
https://smartcontract-it.atlassian.net/browse/DPA-1613?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ