feat: add utils to generate cancel / bypass proposals from timelock schedule proposal (#342)
add utils to generate cancel / bypass proposals from timelock schedule
proposal

## AI Summary
This pull request introduces new methods to the `TimelockProposal`
struct in `timelock_proposal.go` to enhance its functionality. The most
significant changes include the addition of methods to derive
cancellation and bypass proposals from an existing proposal.

Enhancements to `TimelockProposal`:

* Added `DeriveCancellationProposal` method to derive a new proposal
that cancels the current proposal. This method ensures the original
proposal remains unaffected and returns an error if the action is not of
type 'schedule'.
* Added `DeriveBypassProposal` method to derive a new proposal that
bypasses the current proposal. Similar to the cancellation method, it
ensures the original proposal remains unaffected and returns an error if
the action is not of type 'schedule'.