fix(solana): fix accounts added to BypassExecute instruction (#366)
After investigating issues with the RMNRemote curse integration tests,
we realized that we were mishandling the accounts that are passed to the
BypassExecuteBatch instruction. This PR fixes that, by implementing the
following changes:

* add the Program ID to the list of "remaining accounts"
* de-duplicate the accounts in the "remaining accounts" slice, ensuring
the "signer" attribute is unset
* adjust the writable attribute of the non-remaining accounts of the
BypassExecuteBatch instruction itself

In addition, the PR also fixes the predecessor value used when computing
the operation id -- since bypass executions don't respect the
predecessor value, the operation id computed by the onchain Solana
program assumes that the predecessor is always "null".

[DX-610](https://smartcontract-it.atlassian.net/browse/DX-610)

[DX-610]:
https://smartcontract-it.atlassian.net/browse/DX-610?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ