fix(solana): call ClearSignatures in SetRoot on "account in use" error (#374)
This PR enhances the Solana `SetRoot` implementation so as to handle
"Account already in use" errors, which we got several times when
executing changesets in mainnet. In theory, these errors are happening
because Solana's SetRoot does not execute atomically -- instead, it
requires at least 4 instructions:

1. initialize signatures
2. append signatures (at least once)
3. finalize signatures
4. set root

If the first instruction completes successfully but any of the others
fails, it'll leave the signatures PDA in a initialized state but without
the root properly set. This prevents retries at the CI level from
working as the program will then fail, stating the the pda "account is
already in use".

The workaround in this PR is to identify this error and issue a "clear
signatures" instruction -- which should reset the signatures pda -- and
then retry the SetRoot instructions as usual.

The PR also bumps the chainlink-ccip version so as to include a patch
from Terry Tata to a "send transaction" helper that should help deal
with frequent RPC errors found in mainnet.

[DX-768](https://smartcontract-it.atlassian.net/browse/DX-768)

[DX-768]:
https://smartcontract-it.atlassian.net/browse/DX-768?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ