fix(solana): set CU limit as the maximum (1.4M) to avoid CU errors in mainnet (#392)
Due to CU limit problems executing a `solana.SetRoot` call in mainnet,
we're setting the maximum CU limit for all instructions.

This has the downside of increasing the cost of the transations, so
we'll follow up with a better fix that dynamically sets the CU limit
according to the transaction requirements.