fix(solana): fix calculation of AppendInstructionData chunk size
The previous hardcoded constant used as the maximum "chunk size" used in
a Solana "AppendInstructionData" instruction was estimated based on some
tests performed by the ccip team. In practice, however, the value selected
presented a few problems in production, when executing a few larger
proposals.

After investigating how the Solana MCMS programs wrap the instructions -- in
once by the `AppendInstructionData` instruction and then by the `ExecuteBatch`
instruction -- we arrived at a formula that captures the wrapping
logic reasonably well. It does have a few assumptions, though:
  * a single signature will be added to the Solana transaction
  * the transaction won't use an address lookup table
  * the SetComputeUnitLimit instruction will be added to the transaction

Besides these assumptions, it relies on a pre-determined max number of
proofs that will be added to the `ExecuteBatch` transaction. This value
needs to be estimated because the effective number of proofs depends on
the number of operations -- but the number of operations depends on the
size of the chunks, since a smaller chunk will yield more
`AppendInstructionData` calls.

For now, we defined the default number of proofs to be `8`. It's
possible to override the default value by setting the environment
variable `MCMS_SOLANA_NUM_PROOFS`.

The default value of `8` is enough to support proposals that yield up to
256 instructions. Increasing this number allows us to support proposals
with more instructions, at the expense of having a smaller chunk size
and thus needing more instructions to build the operations in the
Timelock.