feat: add all signers helper for config (#402)
Adds a function AllSigner() to get a list of all the nested signers in
the mcms config. This will be useful to determine the matching addresses
when trying to automatically identify the derivation path in the signing
tool for proposals. This assuming that we decide to add the Config as
part of the chain metadata in the proposal.