feat: add a labels field to attach process related metadata to a proposal [DX-1187] (#413)
As part of being able to attach slack related metadata to a proposal, we
propose adding an optional `Metadata` field to attach metadata to a
proposal. This can be used to attach the slack thread ID to a proposal
or in future other metadata (pager duty groups, categories, risk data,
etc) to better categorize proposals or add process related fields to it.