fix: avoid requiring all executors for predecessors calculation (#449)
This pull request addresses an issue in the converters generation logic
for timelock executables. The main change ensures that converters are
generated based on the proposal's chain metadata rather than requiring
all executors, this is specially useful on the mcms cli in CLD where we
need to sometimes load a subset of the chain selectors only so not all
executables are provided.

---------

Signed-off-by: Pablo <pablo.estrada@smartcontract.com>